rag_search_PoC/requirements.txt
streamlit==1.32.0
langchain==0.1.12
langchain-community>=0.0.28
langchain-openai==0.0.8
pypdf==4.0.2
python-docx==1.1.0
sentence-transformers==2.5.1
faiss-cpu==1.7.4
python-dotenv==1.0.1 